+++ client/package.json
@@ -37,7 +37,7 @@
     "prettier": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
     "tailwindcss": "^3.4.4",
-    "vite": "^5.3.1",
+    "vite": "^5.3.3",
     "vite-plugin-svgr": "^4.2.0"
   }
 }
