+++ client/src/components/common/SelectField.jsx
@@ -1,73 +1,44 @@
-import React, { useState } from 'react';
-import Checkbox from '@/components/common/CheckBox';
+import React from 'react';
 
-const SelectField = ({ label, required, options, onChange, selectedOptions, ...props }) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const [searchTerm, setSearchTerm] = useState('');
-
-  const filteredOptions = options.filter(option =>
-    option.label.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  const toggleOption = value => {
-    const newSelectedOptions = selectedOptions.includes(value)
-      ? selectedOptions.filter(v => v !== value)
-      : [...selectedOptions, value];
-    onChange(newSelectedOptions);
-  };
-
-  return (
-    <div className='mb-4 relative'>
-      <button
-        onClick={() => setIsOpen(!isOpen)}
-        className='text-grey border border-gray-300 bg-white focus:ring-1 focus:outline-none focus:ring-mint focus:border-mint_hover font-medium rounded-lg text-base px-5 py-2.5 text-center inline-flex items-center'
-        type='button'>
-        {label}
-        <svg
-          className='w-2.5 h-2.5 ms-3'
-          aria-hidden='true'
-          xmlns='http://www.w3.org/2000/svg'
-          fill='none'
-          viewBox='0 0 10 6'>
-          <path
-            stroke='currentColor'
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            strokeWidth='2'
-            d='m1 1 4 4 4-4'
-          />
-        </svg>
-      </button>
-
-      {isOpen && (
-        <div className='z-10 absolute mt-1 bg-white rounded-lg shadow w-60'>
-          <div className='p-3'>
-            <input
-              type='text'
-              className='block w-full p-2 ps-5 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-1 focus:outline-none focus:ring-mint focus:border-mint_hover'
-              placeholder='검색'
-              value={searchTerm}
-              onChange={e => setSearchTerm(e.target.value)}
-            />
-          </div>
-          <ul className='h-48 px-3 pb-3 overflow-y-auto text-sm text-gray-700'>
-            {filteredOptions.map((option, index) => (
-              <li key={index}>
-                <div className='flex items-center font-medium p-2 ps-2 rounded hover:bg-gray-100'>
-                  <Checkbox
-                    name={`checkbox-item-${index}`}
-                    label={option.label}
-                    checked={selectedOptions.includes(option.value)}
-                    onChange={() => toggleOption(option.value)}
-                  />
-                </div>
-              </li>
-            ))}
-          </ul>
-        </div>
+const SelectField = ({ label, required, options, ...props }) => (
+  <div className='mb-4'>
+    <label className='mb-1 block text-sm font-medium text-gray-700'>
+      {required && (
+        <span
+          className={`block text-sm font-medium text-slate-700 ${required ? "after:ml-0.5 after:text-red-500 after:content-['*']" : ''}`}>
+          {label}
+        </span>
       )}
+    </label>
+
+    <div className='relative'>
+      <select
+        className='w-full appearance-none rounded-md border border-gray-300 bg-white p-2 pr-10 text-sm focus:border-mint focus:outline-none focus:ring-mint'
+        style={{
+          WebkitAppearance: 'none',
+
+          MozAppearance: 'none',
+
+          appearance: 'none',
+        }}
+        {...props}>
+        {options.map((option, index) => (
+          <option key={index} value={option.value}>
+            {option.label}
+          </option>
+        ))}
+      </select>
+
+      <div className='pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700'>
+        <svg className='h-5 w-5 fill-current' viewBox='0 0 20 20'>
+          <path
+            d='M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z'
+            clipRule='evenodd'
+            fillRule='evenodd'></path>
+        </svg>
+      </div>
     </div>
-  );
-};
+  </div>
+);
 
 export default SelectField;
