+++ client/src/components/member/simpConsent/PaymentCard.jsx
@@ -74,6 +74,9 @@ const PaymentCard = () => {
           onBlur={handleBlur}
         />
       </form>
+      <button className='mt-4 w-full rounded-lg border border-teal-400 bg-white py-2 text-sm font-normal text-teal-400 transition-colors hover:bg-teal-50'>
+        카드 인증하기
+      </button>
     </div>
   );
 };
